pyproject.toml
[tool.poetry]
name = "iff-haystag-rag"
version = "0.1.0"
description = ""
authors = ["jhnoor <jamawadi@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
haystack-ai = "^2.9.0"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.3.1"
langchain = "^0.3.17"
langchain-community = "^0.3.16"
langchain-core = "^0.3.33"
langgraph = "^0.2.69"
beautifulsoup4 = "^4.13.1"
faiss-cpu = "^1.10.0"
openai = "^1.61.0"
tiktoken = "^0.8.0"
typing-extensions = "^4.12.2"
langchain-openai = "^0.3.3"
llama-index = "^0.12.15"
llama-index-readers-web = "^0.3.5"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
